update attribution
Given that the original looked like a clone of the comic plugin and you've done so much work, you should take credit for it.

IMHO, given the work you've done, you should clone this repository as it stands and create a new one that the community can actually iterate on since Tyler abandoned the original. This will allow the additional functionality on it that Github doesn't allow for forks.